scripts/pyproject.toml
[tool.poetry]
name = "bert-llm-chat-scripts"
version = "0.1.0"
description = "Ingestion scripts for FAISS index"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
pypdf = "^6.2.0"
sentence-transformers = "^5.1.2"
langchain = "^1.0.5"
faiss-cpu = "^1.12.0"
python-dotenv = "^1.2.1"
langchain-text-splitters = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
